Un-refactor equality checking, since it was wrong for sub.
