+++ evm/src/alu/mod.rs
@@ -1,7 +1,6 @@
 mod add;
 mod mul;
 mod sub;
-mod utils;
 
 pub mod alu_stark;
 pub(crate) mod columns;
