+++ evm/src/alu/sub.rs
@@ -5,8 +5,58 @@ use plonky2::hash::hash_types::RichField;
 use plonky2::iop::ext_target::ExtensionTarget;
 
 use crate::alu::columns;
-use crate::alu::utils;
 use crate::constraint_consumer::{ConstraintConsumer, RecursiveConstraintConsumer};
+
+/// NB: Tests for equality, but only on the assumption that the limbs
+/// in `larger` are all at least as big as those in `smaller`, and
+/// that the limbs in `larger` are at most (LIMB_BITS + 1) bits.
+fn eval_packed_generic_are_equal<P, I, J>(
+    yield_constr: &mut ConstraintConsumer<P>,
+    is_op: P,
+    larger: I,
+    smaller: J,
+) where
+    P: PackedField,
+    I: Iterator<Item = P>,
+    J: Iterator<Item = P>,
+{
+    let mut br = P::ZEROS;
+    for (a, b) in larger.zip(smaller) {
+        // t should be either 0 or 1
+        let t = a - (b + br);
+        yield_constr.constraint(is_op * (t - t * t));
+        br = t;
+    }
+}
+
+fn eval_ext_circuit_are_equal<F, const D: usize, I, J>(
+    builder: &mut plonky2::plonk::circuit_builder::CircuitBuilder<F, D>,
+    yield_constr: &mut RecursiveConstraintConsumer<F, D>,
+    is_op: ExtensionTarget<D>,
+    larger: I,
+    smaller: J,
+) where
+    F: RichField + Extendable<D>,
+    I: Iterator<Item = ExtensionTarget<D>>,
+    J: Iterator<Item = ExtensionTarget<D>>,
+{
+    let mut br = builder.zero_extension();
+    for (a, b) in larger.zip(smaller) {
+        // t0 = b + br
+        let t0 = builder.add_extension(b, br);
+        // t  = a - t0
+        let t = builder.sub_extension(a, t0);
+        // t1 = t * t
+        let t1 = builder.mul_extension(t, t);
+        // t2 = t1 - t
+        let t2 = builder.sub_extension(t1, t);
+
+        let filtered_limb_constraint = builder.mul_extension(is_op, t2);
+        yield_constr.constraint(builder, filtered_limb_constraint);
+
+        br = t;
+    }
+}
 
 pub fn generate<F: RichField>(lv: &mut [F; columns::NUM_ALU_COLUMNS]) {
     let input0_limbs = columns::SUB_INPUT_0.map(|c| lv[c].to_canonical_u64());
@@ -48,7 +98,7 @@ pub fn eval_packed_generic<P: PackedField>(
         .zip(input1_limbs)
         .map(|(a, b)| limb_boundary + a - b);
 
-    utils::eval_packed_generic_are_equal(yield_constr, is_sub, output_computed, output_limbs);
+    eval_packed_generic_are_equal(yield_constr, is_sub, output_computed, output_limbs);
 }
 
 #[allow(clippy::needless_collect)]
@@ -73,7 +123,7 @@ pub fn eval_ext_circuit<F: RichField + Extendable<D>, const D: usize>(
         })
         .collect::<Vec<ExtensionTarget<D>>>();
 
-    utils::eval_ext_circuit_are_equal(
+    eval_ext_circuit_are_equal(
         builder,
         yield_constr,
         is_sub,
